Fixed email:profile:performance job

+++ app/Console/Commands/EmailProfilePerformance.php
@@ -39,13 +39,14 @@ class EmailProfilePerformance extends Command
 
         $profiles = Profile::where('employee', '=', true)->get();
 
-        $daysAgo = $this->argument('daysAgo');
-        $unixNow = (new \DateTime())->format('U');
-        $unixAgo = $unixNow - (int)$daysAgo * 24 * 60 * 60;
+        // Set time range
+        $daysAgo = (int) $this->argument('daysAgo');
+        $unixNow = (int) (new \DateTime())->format('U');
+        $unixAgo = $unixNow - $daysAgo * 24 * 60 * 60;
 
         $forAccountants = $this->option('accountants');
 
-        //if option is passed set date range from 1st day of current month until last day of current month
+        // If option is passed set date range from 1st day of current month until last day of current month
         if ($forAccountants) {
             $unixNow = Carbon::now()->endOfMonth()->format('U');
             $unixAgo = Carbon::now()->firstOfMonth()->format('U');
@@ -63,7 +64,7 @@ class EmailProfilePerformance extends Command
             $data['fromDate'] = \DateTime::createFromFormat('U', $unixAgo)->format('Y-m-d');
             $data['toDate'] = \DateTime::createFromFormat('U', $unixNow)->format('Y-m-d');
 
-            //if option is not passed send mail to each profile
+            // If option is not passed send mail to each profile
             if ($forAccountants === false) {
                 $view = 'emails.profile.performance';
                 $subject = Config::get('mail.private_mail_subject');
@@ -77,7 +78,7 @@ class EmailProfilePerformance extends Command
 
         $overviewRecipients = Profile::where('admin', '=', true)->get();
 
-        //if option is passed get all admins and profiles with accountant role
+        // If option is passed get all admins and profiles with accountant role
         if ($forAccountants) {
             $overviewRecipients = Profile::where('admin', '=', true)
                 ->orWhere('role', '=', 'accountant')
